Merge requirements files
These were previously separate to enforce the idea that the scoring
logic shouldn't depend on srcomp. However since we want to allow
for the scorer UI to be defined in the compstate and to have tests,
it's impractical to approach it this way. Merging these does allow
some risk of an unexpected dependency, however it's probably a
small enough risk not to be troublesome.